requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1
python-dotenv>=1.0
pillow>=10.0
openpyxl>=3.1
pandas>=2.2
reportlab>=4.0
whitenoise>=6.6
django-crispy-forms>=2.1
crispy-tailwind>=1.0
django-extensions>=3.2
requests>=2.32
PyMySQL>=1.1
